StationController wip

+++ src/main/java/org/fffd/l23o6/dao/RouteDao.java
@@ -3,6 +3,6 @@ package org.fffd.l23o6.dao;
 import org.fffd.l23o6.pojo.entity.RouteEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RouteDao extends JpaRepository<RouteEntity, Integer>{
+public interface RouteDao extends JpaRepository<RouteEntity, Long>{
     
 }
